Winforms: • Progress Indicator added. • "Cancel" in the editor now won't prompt to save the recording anymore.
WPF:
• Started the Webcam Capture algorithms and the UI.
• LogWritter class added.
• Few bugfixes with the Numeric fields.
• UI tweaks with the ComboBox.
•

+++ GifRecorder/Pages/Resize.cs
@@ -86,7 +86,7 @@ namespace ScreenToGif.Pages
         /// <param name="a">Size a</param>
         /// <param name="b">Size b</param>
         /// <returns>The GCD number.</returns>
-        static int Gcd(int a, int b)
+        static double Gcd(int a, int b)
         {
             return b == 0 ? a : Gcd(b, a % b);
         }
